Module Sanddb.Database

Sanddb.Database contains the type of the database and it's internal implementation.

module type T = sig ... end

Database.T is the (module) type of the database. In SandDB every database instance will have this type regardles of the serializing method.